Moved the loading of genre data to a new launch activity

+++ app/src/main/java/com/clidev/gamecodex/populargames/model/room/genre/Genre.java
@@ -14,7 +14,7 @@ public class Genre {
     @PrimaryKey
     @SerializedName("id")
     @Expose
-    private Integer id;
+    private Long id;
 
     @SerializedName("name")
     @Expose
@@ -33,16 +33,16 @@ public class Genre {
      * 
      * @param id, name
      */
-    public Genre(Integer id, String name) {
+    public Genre(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
